+++ src/main/java/id/perumdamts/kepegawaian/repositories/kepegawaian/LampiranSkRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 
 public interface LampiranSkRepository extends JpaRepository<LampiranSk, Long>,
         JpaSpecificationExecutor<LampiranSk>,
-        RevisionRepository<LampiranSk, Long, Long> {
+        RevisionRepository<LampiranSk, Long, Integer> {
     List<LampiranSk> findByRefAndRefId(EJenisSk jenisSk, Long id);
 
     List<LampiranSk> findAllByRefId(Long id);
